requirements.txt
behave
requests
allure-behave
allure-python-commons
mysql-connector-python